+++ src/main/java/oshi/jna/platform/mac/IOKit.java
@@ -27,6 +27,7 @@ import com.sun.jna.Pointer;
 import com.sun.jna.Structure;
 import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.LongByReference;
+import com.sun.jna.ptr.PointerByReference;
 
 import oshi.jna.platform.mac.CoreFoundation.CFAllocatorRef;
 import oshi.jna.platform.mac.CoreFoundation.CFArrayRef;
@@ -165,6 +166,8 @@ public interface IOKit extends Library {
 
     CFMutableDictionaryRef IOServiceMatching(String name);
 
+    CFMutableDictionaryRef IOBSDNameMatching(int masterPort, int options, String bsdName);
+
     int IOServiceGetMatchingService(int port, CFMutableDictionaryRef matchingDictionary);
 
     int IOServiceGetMatchingServices(int port, CFMutableDictionaryRef matchingDictionary, IntByReference iterator);
@@ -177,13 +180,18 @@ public interface IOKit extends Library {
 
     int IOIteratorNext(int iterator);
 
+    boolean IOObjectConformsTo(int object, String className);
+
     // Requires OS X 10.5+
     int IOConnectCallStructMethod(int connection, int selector, Structure inputStructure, int structureInputSize,
             Structure outputStructure, IntByReference structureOutputSize);
 
     CFTypeRef IORegistryEntryCreateCFProperty(int entry, CFStringRef key, CFAllocatorRef allocator, int options);
 
-    int IORegistryEntryCreateCFProperties(int entry, CFMutableDictionaryRef serviceProperties, CFAllocatorRef allocator,
+    CFTypeRef IORegistryEntrySearchCFProperty(int entry, String plane, CFStringRef key, CFAllocatorRef allocator,
+            int options);
+
+    int IORegistryEntryCreateCFProperties(int entry, PointerByReference propsPtr, CFAllocatorRef allocator,
             int options);
 
     CFStringRef CFCopyDescription(CFTypeRef type);